+++ cmd/run.go
@@ -95,6 +95,9 @@ func runRunCmd(cmd *cobra.Command, args []string) error {
 			EnableGethTracer:    cfg.EnableGethTracer,
 			FilterTokens:        cfg.TokenFilter,
 			SupportedTokens:     getSupportedTokens(cfg.Network.Network),
+			SuportsSyncing:      cfg.SupportsSyncing,
+			SkipAdminCalls:      false,
+			SupportsPeering:     false,
 		}
 		var err error
 		client, err = optimism.NewClient(cfg.GethURL, cfg.Params, opts)
