+++ optimism/client_test.go
@@ -29,7 +29,6 @@ import (
 	"github.com/inphi/optimism-rosetta/optimism/utilities/artifacts"
 
 	RosettaTypes "github.com/coinbase/rosetta-sdk-go/types"
-	ethereum "github.com/ethereum-optimism/optimism/l2geth"
 	"github.com/ethereum-optimism/optimism/l2geth/common"
 	"github.com/ethereum-optimism/optimism/l2geth/common/hexutil"
 	"github.com/ethereum-optimism/optimism/l2geth/core/types"
@@ -40,144 +39,6 @@ import (
 	"github.com/stretchr/testify/mock"
 	"golang.org/x/sync/semaphore"
 )
-
-func TestStatus_NotReady(t *testing.T) {
-	mockJSONRPC := &mocks.JSONRPC{}
-	mockGraphQL := &mocks.GraphQL{}
-
-	cf, err := newERC20CurrencyFetcher(mockJSONRPC)
-	assert.NoError(t, err)
-
-	c := &Client{
-		c:               mockJSONRPC,
-		g:               mockGraphQL,
-		currencyFetcher: cf,
-		traceSemaphore:  semaphore.NewWeighted(100),
-	}
-
-	ctx := context.Background()
-	mockJSONRPC.On(
-		"CallContext",
-		ctx,
-		mock.Anything,
-		"eth_getBlockByNumber",
-		"latest",
-		false,
-	).Return(
-		nil,
-	).Once()
-
-	block, timestamp, syncStatus, peers, err := c.Status(ctx)
-	assert.Nil(t, block)
-	assert.Equal(t, int64(-1), timestamp)
-	assert.Nil(t, syncStatus)
-	assert.Nil(t, peers)
-	assert.True(t, errors.Is(err, ethereum.NotFound))
-
-	mockJSONRPC.AssertExpectations(t)
-	mockGraphQL.AssertExpectations(t)
-}
-
-func TestStatus_NotSyncing(t *testing.T) {
-	mockJSONRPC := &mocks.JSONRPC{}
-	mockGraphQL := &mocks.GraphQL{}
-
-	c := &Client{
-		c:              mockJSONRPC,
-		g:              mockGraphQL,
-		traceSemaphore: semaphore.NewWeighted(100),
-	}
-
-	ctx := context.Background()
-	mockJSONRPC.On(
-		"CallContext",
-		ctx,
-		mock.Anything,
-		"eth_getBlockByNumber",
-		"latest",
-		false,
-	).Return(
-		nil,
-	).Run(
-		func(args mock.Arguments) {
-			header := args.Get(1).(**types.Header)
-			file, err := ioutil.ReadFile("testdata/basic_header.json")
-			assert.NoError(t, err)
-
-			*header = new(types.Header)
-
-			assert.NoError(t, (*header).UnmarshalJSON(file))
-		},
-	).Once()
-
-	block, timestamp, syncStatus, peers, err := c.Status(ctx)
-	assert.Equal(t, &RosettaTypes.BlockIdentifier{
-		Hash:  "0x48269a339ce1489cff6bab70eff432289c4f490b81dbd00ff1f81c68de06b842",
-		Index: 8916656,
-	}, block)
-	assert.Equal(t, int64(1603225195000), timestamp)
-	assert.Equal(t, &RosettaTypes.SyncStatus{
-		CurrentIndex: RosettaTypes.Int64(8916656),
-		TargetIndex:  RosettaTypes.Int64(8916656),
-	}, syncStatus)
-	assert.Nil(t, peers)
-	assert.NoError(t, err)
-
-	mockJSONRPC.AssertExpectations(t)
-	mockGraphQL.AssertExpectations(t)
-}
-
-func TestStatus_Syncing(t *testing.T) {
-	mockJSONRPC := &mocks.JSONRPC{}
-	mockGraphQL := &mocks.GraphQL{}
-	cf, err := newERC20CurrencyFetcher(mockJSONRPC)
-	assert.NoError(t, err)
-
-	c := &Client{
-		c:               mockJSONRPC,
-		g:               mockGraphQL,
-		currencyFetcher: cf,
-		traceSemaphore:  semaphore.NewWeighted(100),
-	}
-
-	ctx := context.Background()
-	mockJSONRPC.On(
-		"CallContext",
-		ctx,
-		mock.Anything,
-		"eth_getBlockByNumber",
-		"latest",
-		false,
-	).Return(
-		nil,
-	).Run(
-		func(args mock.Arguments) {
-			header := args.Get(1).(**types.Header)
-			file, err := ioutil.ReadFile("testdata/basic_header.json")
-			assert.NoError(t, err)
-
-			*header = new(types.Header)
-
-			assert.NoError(t, (*header).UnmarshalJSON(file))
-		},
-	).Once()
-
-	block, timestamp, syncStatus, peers, err := c.Status(ctx)
-	assert.Equal(t, &RosettaTypes.BlockIdentifier{
-		Hash:  "0x48269a339ce1489cff6bab70eff432289c4f490b81dbd00ff1f81c68de06b842",
-		Index: 8916656,
-	}, block)
-	assert.Equal(t, int64(1603225195000), timestamp)
-	assert.Equal(t, &RosettaTypes.SyncStatus{
-		CurrentIndex: RosettaTypes.Int64(8916656),
-		TargetIndex:  RosettaTypes.Int64(8916656),
-	}, syncStatus)
-	assert.Nil(t, peers)
-	assert.NoError(t, err)
-
-	mockJSONRPC.AssertExpectations(t)
-	mockGraphQL.AssertExpectations(t)
-}
 
 func TestBalance(t *testing.T) {
 	mockJSONRPC := &mocks.JSONRPC{}
